Add GameMenuView and FriendsListComponent
This commit adds the GameMenuView and FriendsListComponent to the project. The GameMenuView is a new view that will be used for the game menu, and the FriendsListComponent is a new component that displays a list of friends with their status. These additions will enhance the user experience and provide more functionality to the application.